requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
Flask==2.2.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.1.1
numpy==1.21.3
openai==0.19.0
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
pycodestyle==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.2.2
flask-wtf==1.1.1
wtforms==3.0.1
email_validator==1.3.1
flask_sqlalchemy==3.0.3
openai-whisper==20230124
ffmpeg==1.4
torch==1.13.1
ffmpeg-python==0.2.0
more-itertools==9.0.0
transformers==4.19.0